fix the liners
